backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
mysqlclient>=2.2
Pillow>=10.0
python-dotenv>=1.0
